ai_service/requirements.txt
grpcio
grpcio-tools
strands-agents
strands-agents-tools
boto3
pytest
